87 auth redirect fix (#42)
* Fix potential timing issue with attempt to load constellation bootstrap when code url is returned
* Minor Readme changes
* Move constellationInit into authWrapper
* Move app specific code up to sample layer
